requirements.txt
certifi==2019.9.11
chardet==3.0.4
clearbit==0.1.7
Click==7.0
Django==2.2.6
django-crispy-forms==1.7.2
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
Faker==2.0.2
idna==2.8
psycopg2==2.8.3
pyhunter==1.2
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
termcolor==1.1.0
text-unidecode==1.3
urllib3==1.25.6
